refactor(tee): split azure.rs god object into type/impl/contract modules
tee/azure.rs (697 LOC) -> tee/azure/: mod.rs keeps the config + token +
AzureSkrBackend struct + require_env; methods.rs holds the inherent impl
(auth token, network + public-IP + NIC creation, CVM body, wait-for-running,
resource delete); backend.rs holds the TeeBackend trait impl. The inherent
helpers + the http/token_cache fields are widened to pub(crate) so the trait
impl (now a sibling module) can call them; tee-level sidecar helpers reached
via super:: are re-exported from mod.rs. Pure code motion, no behavior
change; max module 373 LOC.
