+++ scripts/.god-object-baseline.txt
@@ -12,7 +12,6 @@
 744 sandbox-runtime/src/scoped_session_auth.rs
 721 ui/src/components/shared/SessionSidebar.tsx
 701 ui/src/lib/stores/chatSessions.ts
-697 sandbox-runtime/src/tee/azure.rs
 682 ui/src/components/shared/InfrastructureModal.tsx
 653 ai-agent-instance-blueprint-lib/src/auto_provision.rs
 627 ui/src/lib/hooks/useCreateDeploy.ts
